requirements.txt
torch==2.0.1
torchvision==0.15.2
pillow==9.5.0
numpy==1.24.3
pandas==2.0.1
scikit-learn==1.2.2
